Avoid pam_lastlog prompts in non-interactive mode
If sudo is configured to include session system-auth:

    $ cat /etc/pam.d/sudo
    ...
    session    include      system-auth

And system-auth is using pam_lastlog plugin:

    $ cat /etc/pam.d/system-auth
    ...
    session     required      pam_lastlog.so showfailed

Running a command in non-interactive mode will add the last login info
in the command stdout:

    $ sudo -n echo "expected-output"
    Last login: Wed Jan  2 13:23:51 EST 2019 on pts/0
    expected-output

This breaks use cases when command output is consumed by another
program. An example real use case is running LVM commands using sudo.
LVM output is built to be consumed by programs and programs assume that
sudo is not modifying the output in any way.

Change pam_open_session to use PAM_SLIENT if running in non-interactive
mode, avoiding this issue.